+++ ZaloPay/app/src/main/java/vn/com/vng/zalopay/ui/presenter/ZaloPayPresenter.java
@@ -17,7 +17,7 @@ public interface ZaloPayPresenter<IZaloPayView> extends IPresenter<IZaloPayView>
 
     void getBanners();
 
-    void startServiceWebViewActivity(int appId, String webViewUrl);
+    void startServiceWebViewActivity(long appId, String webViewUrl);
 
     void startBannerCountDownTimer();
