+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/api/entity/UserExistEntity.java
@@ -19,5 +19,5 @@ public class UserExistEntity {
     @SerializedName("zalopayname")
     public String zalopayname;
     @SerializedName("status")
-    public int status;
+    public long status;
 }
