+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/appresources/DownloadAppResourceTask.java
@@ -87,7 +87,7 @@ final class DownloadAppResourceTask {
     }
 
 
-    private String getExternalBundleFolder(int appId) {
+    private String getExternalBundleFolder(long appId) {
         //return String.format(Locale.getDefault(), "%s/modules/%d/app", mBundleRootFolder, appId);
         return ResourceHelper.getPath(appId);
     }
